requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
sqlalchemy==2.0.23  # Use SQLAlchemy 1.x for Python 3.13 compatibility
pydantic==2.5.2  # Use a version compatible with Python 3.13
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.9
requests==2.31.0
togetherai==0.0.3
python-magic==0.4.27
gunicorn==21.2.0